requirement.txt
transformers
torch
numpy
scikit-learn
nltk
python-docx
requests
spacy
fastapi 
uvicorn